requirements.txt
python-dotenv
openai
openai-agents 